extract example_utils relational_utils and candidate_collection from ConditionExampleGenerator
